mlmllm

+++ 0x15-file_io/1-create_file.c
@@ -13,7 +13,7 @@
 int create_file(const char *filename, char *text_content)
 {
 	int file_descriptor;
-	ssize_t number, bytes_read;
+	int number, bytes_read;
 
 	if (filename == NULL)
 		return (-1);
